CMS/requirements/base.txt
backports.shutil-get-terminal-size==1.0.0
configparser==3.5.0
decorator==4.0.10
dj-static==0.0.6
Django==1.9.5
django-allauth==0.27.0
django-appconf==1.0.2
django-compressor==2.1
django-extensions==1.7.3
django-rest-auth==0.8.1
djangorestframework==3.4.6
drf-nested-routers==0.11.1
elasticsearch==5.0.1
enum34==1.1.6
flake8==3.0.4
gunicorn==19.6.0
ipython==5.1.0
ipython-genutils==0.1.0
mccabe==0.5.2
oauthlib==2.0.0
pathlib2==2.1.0
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
psycopg2==2.6.2
ptyprocess==0.5.1
pycodestyle==2.0.0
pyflakes==1.3.0
Pygments==2.1.3
python-dateutil==2.6.0
python-openid==2.2.5
pytz==2017.2
rcssmin==1.0.6
requests==2.11.1
requests-oauthlib==0.6.2
rjsmin==1.0.12
simplegeneric==0.8.1
six==1.10.0
static3==0.7.0
traitlets==4.2.2
urllib3==1.19.1
wcwidth==0.1.7
